% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelling.R
\name{apply_labels}
\alias{apply_labels}
\title{Flexible labelling based on named vector.}
\usage{
apply_labels(data, label.vec = var_labels(), overwrite = TRUE)
}
\arguments{
\item{data}{data set}

\item{label.vec}{vector of named character strings, with names being column
names.}

\item{overwrite}{overwrite existing labels. Default is TRUE}
}
\value{
data.frame/tibble with labels
}
\description{
Uses column name if label is missing.
}
\examples{
mtcars |>
  apply_labels() |>
  sapply(attributes) |>
  unlist()
}
