% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Search for child dependencies in package dependencies}
\usage{
check_dependencies(path = here::here(""), packages = NULL)
}
\arguments{
\item{path}{path to the package or project. Default is the source folder of
the currently active project (`here::here("")`)}

\item{packages}{character string of packages to search for}
}
\value{
list
}
\description{
Search for child dependencies in package dependencies
}
\examples{
# Examples not run
# check_dependencies(packages = c("curl"))
# check_dependencies()|>purrr::list_c() |> unique() |> sort()
}
