% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_microdata.R
\name{check_microdata}
\alias{check_microdata}
\title{Microdata violation evaluation}
\usage{
check_microdata(data, n.min)
}
\arguments{
\item{data}{table to evaluate}

\item{n.min}{count difference threshold (default is 3)}
}
\value{
gt_tbl list object
}
\description{
Evaluates if tables contains microdata. Initially supports
`gtsummary::tbl_summary()` tables. Colors problematic cells by two criteria:
1) contains counts less than a specified threshold (default is three), or 2)
contains counts that are less than the specified threshold from other cells
in the same row.
}
\examples{
# mtcars |> gtsummary::tbl_summary(by = gear) |> check_microdata()
}
