% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_microdata.R
\name{check_microdata.tbl_summary}
\alias{check_microdata.tbl_summary}
\title{Microdata violation evaluation}
\usage{
\method{check_microdata}{tbl_summary}(data, n.min = 3)
}
\arguments{
\item{data}{table to evaluate}

\item{n.min}{count difference threshold (default is 3)}
}
\value{
gt_tbl list object
}
\description{
Microdata violation evaluation
}
