% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combn_perm.R
\name{combn_perm}
\alias{combn_perm}
\title{Extension of base::combn() to get all permutations of group size(s) m}
\usage{
combn_perm(data, m, perm = FALSE, force = FALSE, limit = 1e+06)
}
\arguments{
\item{data}{data}

\item{m}{group size(s). Numeric vector of length 1:length(data).}

\item{perm}{logical to include all group sizes from 1:m. If m is missing
1:length(data).}

\item{force}{logical to force complete combinations}

\item{limit}{warning limit of large number of permutations}
}
\value{
list
}
\description{
A few notes on the math and calculating possible options:
https://math.stackexchange.com/a/2339815
}
\examples{
combn_perm(1:3, 1)
combn_perm(1:4, 2)
combn_perm(1:4, c(1, 2, 4), perm = TRUE)
combn_perm(1:10, c(1, 2, 4), perm = FALSE)
combn_perm(letters[1:4], perm = TRUE)
rownames(mtcars) |>
  combn_perm(2) |> length()
\dontrun{
rownames(mtcars) |>
  combn_perm(2,limit=100)
}
}
