% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na_vars.R
\name{fill_na_vars}
\alias{fill_na_vars}
\title{Combine variables to fillmissing values}
\usage{
fill_na_vars(...)
}
\arguments{
\item{...}{a prioritised selection of vectors/variables of same length}
}
\value{
vector
}
\description{
Export the first non-NA value for each cell in rowwise fashion
}
\examples{
fill_na_vars(c(1:3,NA,NA),c(2,NA,3:5))
}
