% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelling.R
\name{get_match}
\alias{get_match}
\title{Retrieved matched values}
\usage{
get_match(data, target, named = FALSE)
}
\arguments{
\item{data}{data}

\item{target}{target to match against and subset from}

\item{named}{keep names. Default is FALSE.}
}
\value{
character vector
}
\description{
Retrieved matched values
}
