% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_write.R
\name{gtsummary_write}
\alias{gtsummary_write}
\title{Small wrapper to export table with 'gt'}
\usage{
gtsummary_write(data, path, ...)
}
\arguments{
\item{data}{gtsummary object}

\item{path}{output path. Format follows file extension in path.}

\item{...}{passed on to gt::gtsave}
}
\description{
Small wrapper to export table with 'gt'
}
\examples{
# gtsummary::trial |> gtsummary::tbl_summary() |> gtsummary2docx()
}
