% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shinylive.R
\name{launch_shinylive}
\alias{launch_shinylive}
\title{Generate and launch shinylive instance}
\usage{
launch_shinylive(app.dir = here::here("app/"), dest.dir = here::here("docs/"))
}
\arguments{
\item{app.dir}{dir to app}

\item{dest.dir}{destination dir}
}
\value{
shinylive app
}
\description{
Generate and launch shinylive instance
}
