% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_tics.R
\name{log_tics}
\alias{log_tics}
\title{Create symmetrical tics for a log scale based on sequence}
\usage{
log_tics(data, dec = 2)
}
\arguments{
\item{data}{numeric vector}

\item{dec}{decimals}
}
\value{
numeric vector
}
\description{
Create symmetrical tics for a log scale based on sequence
}
\examples{
log_tics(c(.5,1,5))
}
