% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scripts.R
\name{merge_scripts}
\alias{merge_scripts}
\title{Merge scripts to single file. Handy for shinylive apps}
\usage{
merge_scripts(files, dest, strip.roxygen = TRUE)
}
\arguments{
\item{files}{files to merge. Use list.files().}

\item{dest}{destination path}

\item{strip.roxygen}{optionally remove roxygen comment lines}
}
\description{
Merge scripts to single file. Handy for shinylive apps
}
