% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscure_number.R
\name{obscure_number}
\alias{obscure_number}
\title{Obscure or offset numeric vector by a random number within a given interval}
\usage{
obscure_number(data, dif = 3)
}
\arguments{
\item{data}{numeric vector}

\item{dif}{difference to offset original}
}
\value{
numeric vector
}
\description{
Obscure or offset numeric vector by a random number within a given interval
}
\examples{
obscure_number(1:10)
# obscure_number(letters[1:10])
# obscure_number(1:10,1:2)
}
