% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_seq.R
\name{split_seq}
\alias{split_seq}
\title{Splits sequence in n groups of equal size or groups of max l}
\usage{
split_seq(sequence, n = NULL, l = NULL, split.labels = NULL)
}
\arguments{
\item{sequence}{vector}

\item{n}{number of groups. Ignored if l is specified.}

\item{l}{max length of groups}

\item{split.labels}{optional label for groups}
}
\value{
list
}
\description{
Splits sequence in n groups of equal size or groups of max l
}
\examples{
split_seq(1:8, l = 3)
}
