% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelling.R
\name{var_labels}
\alias{var_labels}
\title{Functionalised character vector of all labels}
\usage{
var_labels()
}
\value{
vector of named labels
}
\description{
This just serves as a template.
}
\examples{
var_labels()
}
